two digits precision for distances and monetary values

+++ app/driver-offer-bpp/src/Domain/Action/UI/Location.hs
@@ -46,6 +46,6 @@ getLocation rideId = do
     DrLoc.findById driver.id
       >>= fromMaybeM LocationNotFound
   let lastUpdate = currLocation.updatedAt
-  let totalDistance = ride.traveledDistance.getHighPrecMeters
+  let totalDistance = realToFrac ride.traveledDistance.getHighPrecMeters
       currPoint = GoogleMaps.getCoordinates currLocation
   return $ GetLocationRes {..}
